Create A frontend project for a Tour and Travel website built with HTML, CSS, and JavaScript. It highlights exciting travel packages to destinations like Rajasthan, Manali, and Shillong. The site is designed to be user-friendly, with features like a search bar, a contact form, and an easy-to-navigate menu, making it simple for travelers to explore and plan their trips."
A frontend project for a Tour and Travel website built with HTML, CSS, and JavaScript. It highlights exciting travel packages to destinations like Rajasthan, Manali, and Shillong. The site is designed to be user-friendly, with features like a search bar, a contact form, and an easy-to-navigate menu, making it simple for travelers to explore and plan their trips.